requirements.txt
flask
flask_login
flask_wtf
flask_sqlalchemy
argon2-cffi
